Integration and unbreaking of MSE support changes. There are still several major issues -       MSE cards look decent, but there is virtually no text prettification.       Split cards (bsides) are not supported at all.       There is no abuse of the notes field to dump additional data yet.

+++ decode.py
@@ -1,11 +1,8 @@
-#!c:/Python27/python.exe -u
 #!/usr/bin/env python
 import sys
 import os
 import zipfile
 import shutil
-
-#to use: py decode.py homebrew.txt homepretty.txt --norarity -v -mse in mtgencode folder.
 
 libdir = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'lib')
 sys.path.append(libdir)
@@ -19,7 +16,8 @@ def exclude_sets(cardset):
     return cardset == 'Unglued' or cardset == 'Unhinged' or cardset == 'Celebration'
 
 def main(fname, oname = None, verbose = True, 
-         gatherer = False, for_forum = False, creativity = False, norarity = False, for_mse = False):
+         gatherer = False, for_forum = False, for_mse = False,
+         creativity = False, norarity = False):
     cards = []
     valid = 0
     invalid = 0
@@ -113,11 +111,18 @@ def main(fname, oname = None, verbose = True,
 
     def writecards(writer):
         if for_mse:
-            # have to prepend a massive chunk.
+            # have to prepend a massive chunk of formatting info
             writer.write(utils.mse_prepend)
         for card in cards:
-            writer.write((card.format(gatherer = gatherer, for_forum = for_forum, for_mse = for_mse)))
-            if creativity and not for_mse: # this won't end well if mse mode is enabled.
+            if for_mse:
+                 writer.write(card.to_mse().encode('utf-8'))
+            else:
+                writer.write(card.format(gatherer = gatherer, 
+                                         for_forum = for_forum).encode('utf-8'))
+
+            if creativity:
+                if for_mse:
+                    writer.write('\tnotes:\n\t\t'.encode('utf-8'))
                 writer.write('~~ closest cards ~~\n'.encode('utf-8'))
                 nearest = cbow.nearest(card)
                 for dist, cardname in nearest:
@@ -133,8 +138,10 @@ def main(fname, oname = None, verbose = True,
                         cardname = '[card]' + cardname + '[/card]'
                     writer.write((cardname + ': ' + str(dist) + '\n').encode('utf-8'))
             writer.write('\n'.encode('utf-8'))
+
         if for_mse:
-            writer.write('version control:\n\ttype: none\napprentice code: ') # have to append some junk at the end of file.
+            # more formatting info
+            writer.write('version control:\n\ttype: none\napprentice code: ')
 
     if oname:
         if verbose:
@@ -142,16 +149,21 @@ def main(fname, oname = None, verbose = True,
         with open(oname, 'w') as ofile:
             writecards(ofile)
         if for_mse:
-            shutil.copyfile(oname, 'set') # copy whatever output file is produced, name the copy 'set' (yes, no extension).
-            zf = zipfile.ZipFile(oname+'.mse-set', mode='w') # use the freaky mse extension instead of zip.
-            try:
-                zf.write('set') # zip up the set file into oname.mse-set.
-            finally:
-                print 'Made an MSE set file called ' + oname + '.mse-set.'
-                zf.close()
-                os.remove('set') # the set file is useless outside the .mse-set, delete it.
-
-
+            # Copy whatever output file is produced, name the copy 'set' (yes, no extension).
+            if os.path.isfile('set'):
+                print 'ERROR: tried to overwrite existing file "set" - aborting.'
+                return
+            shutil.copyfile(oname, 'set')
+            # Use the freaky mse extension instead of zip.
+            with zipfile.ZipFile(oname+'.mse-set', mode='w') as zf:
+                try:
+                    # Zip up the set file into oname.mse-set.
+                    zf.write('set') 
+                finally:
+                    if verbose:
+                        print 'Made an MSE set file called ' + oname + '.mse-set.'
+                    # The set file is useless outside the .mse-set, delete it.
+                    os.remove('set') 
     else:
         writecards(sys.stdout)
         sys.stdout.flush()
@@ -179,6 +191,6 @@ if __name__ == '__main__':
     
     args = parser.parse_args()
     main(args.infile, args.outfile, verbose = args.verbose, 
-         gatherer = args.gatherer, for_forum = args.forum, creativity = args.creativity,
-         norarity = args.norarity, for_mse = args.mse)
+         gatherer = args.gatherer, for_forum = args.forum, for_mse = args.mse,
+         creativity = args.creativity, norarity = args.norarity)
     exit(0)
